Move the USER_AGENT to a more generic const file.
