+++ python/semantic_kernel/connectors/ai/open_ai/const.py
@@ -3,4 +3,3 @@
 from typing import Final
 
 DEFAULT_AZURE_API_VERSION: Final[str] = "2024-02-01"
-USER_AGENT: Final[str] = "User-Agent"
