+++ python/semantic_kernel/const.py
@@ -4,3 +4,4 @@ from typing import Final
 
 METADATA_EXCEPTION_KEY: Final[str] = "exception"
 DEFAULT_SERVICE_NAME: Final[str] = "default"
+USER_AGENT: Final[str] = "User-Agent"
